🎨 Moved Animations into separate file.
